Crate imagequant

https://pngquant.org/lib/

Converts RGBA images to 8-bit with alpha channel.

See examples/ directory for example code.

Modules

attr 🔒

blur 🔒

error 🔒

hist 🔒

image 🔒

kmeans 🔒

mediancut 🔒

nearest 🔒

pal 🔒

quant 🔒

rayoff 🔒

remap 🔒

rows 🔒

seacow 🔒

Structs

Attributes

Starting point and settings for the quantization process

CacheLineAlign 🔒

Histogram

Generate one shared palette for multiple images

HistogramEntry

Number of pixels in a given color for Histogram::add_colors()

Image

Describes image dimensions and pixels for the library

OrdFloat 🔒

Rust is too conservative about sorting floats. This library uses only finite values, so they’re sortable.

QuantizationResult

Remapping step, computed from Attributes::quantize()

Enums

ControlFlow

Result of callback in Attributes::set_progress_callback

Error

Error codes

Constants

LIQ_HIGH_MEMORY_LIMIT 🔒

Traits

PushInCapacity 🔒

Functions

new

Start here: creates new handle for library configuration

Type Aliases

RGBA

8-bit RGBA in sRGB. This is the only color format publicly used by the library.