imagequant::attr

Struct Attributes

pub struct Attributes {

    pub(crate) max_colors: u16,
    target_mse: f64,
    max_mse: Option<f64>,
    kmeans_iteration_limit: f64,
    kmeans_iterations: u16,
    feedback_loop_trials: u16,
    pub(crate) max_histogram_entries: u32,
    min_posterization_output: u8,
    min_posterization_input: u8,
    pub(crate) last_index_transparent: bool,
    pub(crate) use_contrast_maps: bool,
    pub(crate) single_threaded_dithering: bool,
    pub(crate) use_dither_map: DitherMapMode,
    speed: u8,
    pub(crate) progress_stage1: u8,
    pub(crate) progress_stage2: u8,
    pub(crate) progress_stage3: u8,
    progress_callback: Option<Arc<dyn Fn(f32) -> ControlFlow + Send + Sync>>,
    log_callback: Option<Arc<dyn Fn(&Attributes, &str) + Send + Sync>>,
    log_flush_callback: Option<Arc<dyn Fn(&Attributes) + Send + Sync>>,

}

Starting point and settings for the quantization process

Fields
max_colors: u16target_mse: f64max_mse: Option<f64>kmeans_iteration_limit: f64kmeans_iterations: u16feedback_loop_trials: u16max_histogram_entries: u32min_posterization_output: u8min_posterization_input: u8last_index_transparent: booluse_contrast_maps: boolsingle_threaded_dithering: booluse_dither_map: DitherMapModespeed: u8progress_stage1: u8progress_stage2: u8progress_stage3: u8progress_callback: Option<Arc<dyn Fn(f32) -> ControlFlow + Send + Sync>>log_callback: Option<Arc<dyn Fn(&Attributes, &str) + Send + Sync>>log_flush_callback: Option<Arc<dyn Fn(&Attributes) + Send + Sync>>
Implementations

impl Attributes

pub fn new() -> Self

New handle for library configuration

See also Attributes::new_image()

pub fn new_image<VecRGBA>( &self, pixels: VecRGBA, width: usize, height: usize, gamma: f64, ) -> Result<Image<'static>, Error>
where VecRGBA: Into<Box<[RGBA]>>,

Make an image from RGBA pixels.

The pixels argument can be Vec<RGBA>, or Box<[RGBA]> or &[RGBA]. See Attributes::new_image_borrowed for a non-copying alternative.

Use 0.0 for gamma if the image is sRGB (most images are).

pub fn quantize( &self, image: &mut Image<'_>, ) -> Result<QuantizationResult, Error>

Generate palette for the image

pub fn set_max_colors(&mut self, colors: u32) -> Result<(), Error>

It’s better to use set_quality()

pub fn set_quality(&mut self, minimum: u8, target: u8) -> Result<(), Error>

Range 0-100, roughly like JPEG.

If the minimum quality can’t be met, the quantization will be aborted with an error.

Default is min 0, max 100, which means best effort, and never aborts the process.

If max is less than 100, the library will try to use fewer colors. Images with fewer colors are not always smaller, due to increased dithering it causes.

pub fn set_speed(&mut self, value: i32) -> Result<(), Error>

1-10.

Faster speeds generate images of lower quality, but may be useful for real-time generation of images.

The default is 4.

pub fn set_min_posterization(&mut self, value: u8) -> Result<(), Error>

Number of least significant bits to ignore.

Useful for generating palettes for VGA, 15-bit textures, or other retro platforms.

pub fn min_posterization(&self) -> u8

Returns number of bits of precision truncated

pub fn speed(&self) -> u32

Return currently set speed/quality trade-off setting

pub fn max_colors(&self) -> u32

Return max number of colors set

pub fn quality(&self) -> (u8, u8)

Reads values set with set_quality

pub fn new_image_borrowed<'pixels>( &self, bitmap: &'pixels [RGBA], width: usize, height: usize, gamma: f64, ) -> Result<Image<'pixels>, Error>

Describe dimensions of a slice of RGBA pixels

Use 0.0 for gamma if the image is sRGB (most images are).

pub fn new_image_stride<VecRGBA>( &self, pixels: VecRGBA, width: usize, height: usize, stride: usize, gamma: f64, ) -> Result<Image<'static>, Error>
where VecRGBA: Into<Box<[RGBA]>>,

Like new_image_stride_borrowed, but makes a copy of the pixels.

The pixels argument can be Vec<RGBA>, or Box<[RGBA]> or &[RGBA].

pub fn set_log_callback<F: Fn(&Self, &str) + Send + Sync + 'static>( &mut self, callback: F, )

Set callback function to be called every time the library wants to print a message.

To share data with the callback, use Arc or Atomic* types and move || closures.

pub fn set_log_flush_callback<F: Fn(&Self) + Send + Sync + 'static>( &mut self, callback: F, )

Callback for flushing output (if you buffer messages, that’s the time to flush those buffers)

pub fn set_progress_callback<F: Fn(f32) -> ControlFlow + Send + Sync + 'static>( &mut self, callback: F, )

Set callback function to be called every time the library makes a progress. It can be used to cancel operation early.

To share data with the callback, use Arc or Atomic* types and move || closures.

pub fn set_last_index_transparent(&mut self, is_last: bool)

Move transparent color to the last entry in the palette

This is less efficient for PNG, but required by some broken software

pub(crate) fn progress(&self, percent: f32) -> bool

pub(crate) fn verbose_print(&self, msg: impl AsRef<str>)

pub(crate) fn verbose_printf_flush(&self)

pub(crate) fn feedback_loop_trials(&self, hist_items: usize) -> u16

pub(crate) fn target_mse( &self, hist_items_len: usize, ) -> (Option<f64>, f64, bool)

max_mse, target_mse, user asked for perfect quality

pub(crate) fn kmeans_iterations( &self, hist_items_len: usize, palette_error_is_known: bool, ) -> (u16, f64)

returns iterations, iteration_limit

pub(crate) fn posterize_bits(&self) -> u8