imagequant

Module hist

Structs

Cluster 🔒

HashColor 🔒

ignores the index

HistItem 🔒

Histogram

Generate one shared palette for multiple images

HistogramEntry

Number of pixels in a given color for Histogram::add_colors()

HistogramInternal 🔒

TempHistItem 🔒

U32Hasher 🔒

Constants

LIQ_MAXCLUSTER

Clusters form initial boxes for quantization, to ensure extreme colors are better represented

Type Aliases

FixedColorsSet 🔒

Unions

HistSortTmp

RGBAInt 🔒