imagequant::kmeans

Struct Kmeans

pub(crate) struct Kmeans {
    averages: Vec<ColorAvg>,
    weighed_diff_sum: f64,
}

K-Means iteration: new palette color is computed from weighted average of colors that map best to that palette entry.

Fields
averages: Vec<ColorAvg>weighed_diff_sum: f64
Implementations

impl Kmeans

pub fn new(pal_len: usize) -> Result<Self, Error>

pub fn update_color(&mut self, px: f_pixel, value: f32, matched: u8)

pub fn finalize(self, palette: &mut PalF) -> f64

pub(crate) fn iteration( hist: &mut HistogramInternal, palette: &mut PalF, adjust_weight: bool, ) -> Result<f64, Error>

fn iterate_batch( &mut self, batch: &mut [HistItem], n: &Nearest<'_>, colors: &[f_pixel], adjust_weight: bool, )

pub fn merge(self, new: Self) -> Self

pub fn try_merge<E>( old: Result<Self, E>, new: Result<Self, E>, ) -> Result<Self, E>