imagequant

Module mediancut

Structs

MBox 🔒

MedianCutter 🔒

Functions

hist_item_sort_half 🔒

sorts the slice to make the sum of weights lower than weight_half_sum one side, returns index of the edge between halfvar parts of the set

mediancut 🔒

qsort_partition 🔒

qsort_pivot 🔒

weighed_average_color 🔒