imagequant

Module pal

Structs

PalF 🔒

A palette of premultiplied ARGB 4xf32 colors in internal gamma

PalPop 🔒

To keep the data dense, is_fixed is stuffed into the sign bit

Palette

Not used in the Rust API. RGBA colors obtained from QuantizationResult

f_pixel

4xf32 color using internal gamma.

Constants

INTERNAL_GAMMA

LIQ_WEIGHT_A

LIQ_WEIGHT_B

LIQ_WEIGHT_G

LIQ_WEIGHT_MSE

This is a fudge factor - reminder that colors are not in 0..1 range any more

LIQ_WEIGHT_R

MAX_COLORS 🔒

Palettes are stored on the stack, and really large ones will cause stack overflows

MAX_TRANSP_A

MIN_OPAQUE_A

Functions

gamma_lut

posterize_channel 🔒

Type Aliases

ARGBF

PalIndex

PalIndexRemap

This could be increased to support > 256 colors in remapping too

PalLen

RGBA

8-bit RGBA in sRGB. This is the only color format publicly used by the library.