imagequant::quant

Struct QuantizationResult

pub struct QuantizationResult {
    remapped: Option<Box<Remapped>>,
    pub(crate) palette: PalF,
    progress_callback: Option<Box<dyn Fn(f32) -> ControlFlow + Send + Sync>>,
    pub(crate) int_palette: Palette,
    pub(crate) dither_level: f32,
    pub(crate) gamma: f64,
    pub(crate) palette_error: Option<f64>,
    pub(crate) min_posterization_output: u8,
    pub(crate) use_dither_map: DitherMapMode,
    pub(crate) single_threaded_dithering: bool,
}

Remapping step, computed from Attributes::quantize()

Fields
remapped: Option<Box<Remapped>>palette: PalFprogress_callback: Option<Box<dyn Fn(f32) -> ControlFlow + Send + Sync>>int_palette: Palettedither_level: f32gamma: f64palette_error: Option<f64>min_posterization_output: u8use_dither_map: DitherMapModesingle_threaded_dithering: bool
Implementations

impl QuantizationResult

pub(crate) fn new( attr: &Attributes, hist: HistogramInternal, freeze_result_colors: bool, gamma: f64, ) -> Result<Self, Error>

pub(crate) fn write_remapped_image_rows_internal( &mut self, image: &mut Image<'_>, output_pixels: RowBitmapMut<'_, MaybeUninit<u8>>, ) -> Result<(), Error>

fn optionally_generate_dither_map( use_dither_map: DitherMapMode, image: &mut Image<'_>, uses_background: bool, output_pixels: &mut RowBitmapMut<'_, MaybeUninit<u8>>, palette: &mut PalF, ) -> Result<Option<f64>, Error>

pub fn set_dithering_level(&mut self, value: f32) -> Result<(), Error>

Set to 1.0 to get nice smooth image

pub fn set_output_gamma(&mut self, value: f64) -> Result<(), Error>

The default is sRGB gamma (~1/2.2)

pub fn output_gamma(&self) -> f64

Approximate gamma correction value used for the output

Colors are converted from input gamma to this gamma

pub fn quantization_quality(&self) -> Option<u8>

Number 0-100 guessing how nice the input image will look if remapped to this palette

pub fn quantization_error(&self) -> Option<f64>

Approximate mean square error of the palette

pub fn remapping_error(&self) -> Option<f64>

Approximate mean square error of the palette used for the most recent remapping

pub fn remapping_quality(&self) -> Option<u8>

Palette remapping error mapped back to 0-100 scale, same as the scale in Attributes::set_quality()

pub fn palette(&mut self) -> &[RGBA]

The final palette

It’s slighly better if you get palette from the remapped() call instead

pub(crate) fn int_palette(&mut self) -> &Palette

pub fn set_progress_callback<F: Fn(f32) -> ControlFlow + Sync + Send + 'static>( &mut self, callback: F, )

Callback called at various point of processing, which gets percentage of progress done, and can return ControlFlow::Break to abort further processing

pub(crate) fn remap_progress(&self, percent: f32) -> bool

pub fn remapped( &mut self, image: &mut Image<'_>, ) -> Result<(Vec<RGBA>, Vec<u8>), Error>

Remap image into a palette + indices.

Returns the palette and a 1-byte-per-pixel uncompressed bitmap

pub fn remap_into_vec( &mut self, image: &mut Image<'_>, buf: &mut Vec<u8>, ) -> Result<Vec<RGBA>, Error>

Remap image into an existing buffer. Use remapped() if you don’t have a pre-allocated buffer to reuse.

Writes 1-byte-per-pixel uncompressed bitmap into the Vec.

Returns the palette.

pub fn remap_into( &mut self, image: &mut Image<'_>, output_buf: &mut [MaybeUninit<u8>], ) -> Result<(), Error>

Remap image into an existing buffer.

This is a low-level call for use when existing memory has to be reused. Use remapped() or remap_into_vec() if possible.

Writes 1-byte-per-pixel uncompressed bitmap into the pre-allocated buffer.

You should call palette() after this call, but not before it, because remapping refines the palette.

pub fn palette_vec(&mut self) -> Vec<RGBA>

The final palette, copied.

It’s slighly better if you get palette from the remapped() call instead

pub fn palette_len(&mut self) -> usize

Expected length of the palette

Reads the length without finalizing the colors

pub fn from_palette( attr: &Attributes, palette: &[RGBA], gamma: f64, ) -> Result<Self, Error>

Shortcut for making Histogram with add_fixed_color

Set gamma to 0. for sRGB colors.

pub fn dithering_level(&self) -> f32

Getter for the value set in [set_dithering_level]