imagequant

Struct Histogram

pub struct Histogram {
    gamma: Option<f64>,
    fixed_colors: HashSet<HashColor, U32Hasher>,
    hashmap: HashMap<u32, (u32, RGBA), U32Hasher>,
    posterize_bits: u8,
    max_histogram_entries: u32,
}

Generate one shared palette for multiple images

If you’re converting one image at a time, see Attributes::new_image instead

Fields
gamma: Option<f64>fixed_colors: HashSet<HashColor, U32Hasher>hashmap: HashMap<u32, (u32, RGBA), U32Hasher>

The key is the RGBA cast to u32 The value is a (boosted) count or 0 if it’s a fixed color

posterize_bits: u8max_histogram_entries: u32
Implementations

impl Histogram

pub fn new(attr: &Attributes) -> Self

Creates histogram object that will be used to collect color statistics from multiple images.

All options should be set on attr before the histogram object is created. Options changed later may not have effect.

pub fn add_image( &mut self, attr: &Attributes, image: &mut Image<'_>, ) -> Result<(), Error>

“Learns” colors from the image, which will be later used to generate the palette.

Fixed colors added to the image are also added to the histogram. If the total number of fixed colors exceeds 256, this function will fail with LIQ_BUFFER_TOO_SMALL.

pub fn add_colors( &mut self, entries: &[HistogramEntry], gamma: f64, ) -> Result<(), Error>

Alternative to add_image(). Intead of counting colors in an image, it directly takes an array of colors and their counts.

This function is only useful if you already have a histogram of the image from another source.

The gamma may be 0 to mean sRGB. All calls to add_colors and add_fixed_color should use the same gamma value.

pub fn add_fixed_color(&mut self, rgba: RGBA, gamma: f64) -> Result<(), Error>

Add a color guaranteed to be in the final palette

The gamma may be 0 to mean sRGB. All calls to add_colors and add_fixed_color should use the same gamma value.

pub fn quantize( &mut self, attr: &Attributes, ) -> Result<QuantizationResult, Error>

Generate palette for all images/colors added to the histogram.

Palette generated using this function won’t be improved during remapping. If you’re generating palette for only one image, it’s better not to use the Histogram.

pub(crate) fn quantize_internal( &mut self, attr: &Attributes, freeze_result_colors: bool, ) -> Result<QuantizationResult, Error>

fn add_color(&mut self, rgba: RGBA, boost: u32)

fn reserve(&mut self, entries: usize)

const fn posterize_mask(&self) -> u32

fn init_posterize_bits(&mut self, posterize_bits: u8)

optionallys et

pub(crate) fn add_pixel_rows( &mut self, image: &DynamicRows<'_, '_>, importance_map: Option<&[u8]>, posterize_bits: u8, ) -> Result<(), Error>

pub(crate) fn finalize_builder( &mut self, gamma: f64, ) -> Result<HistogramInternal, Error>